Move tests from bus/mod.rs to tests/async_tests.rs
